requirements.txt
streamlit>=1.30.0
pandas>=1.5.0
plotly>=5.18.0
yfinance<0.2.40
requests>=2.31.0
numpy>=1.24.0
python-dotenv
google-generativeai
pandas_market_calendars
sqlalchemy>=2.0.0
pg8000>=1.30.0
toml>=0.10.0
